Update README.md
I have added a change to the postgres commands as the one before didn't work for me. I also added INTERNAL_FHIR_REQUEST_URL to the environment variables. Lastly, I changed the docker run commands slightly to take care of the host ip issue and so that the containers can talk to one another.